requirements.txt
streamlit
pandas
matplotlib
python-dotenv
snowflake-connector-python
snowflake-snowpark-python
requests
plotly
